controller-client/requirements.txt
websocket-client==0.34.0
six==1.10.0